Fix backspace on non-ascii characters
Turns out that we previously relied on the pty implementation to only
send backspace sequences (more specifically backspace-space-backspace
sequences) for correct behaviour when deleting more characters that we
put in:
It seems like that the (Linux) pty implementation counts how man bytes
were input and only sends so many backspace sequences. Non-ascii
characters such as ö are sent as two bytes which counts as two. Many
backspaces after entering such a character would result in the cursor
being placed outside (on the left) of the window. We fix this by
clamping the cursor position back to 0 (and by incorporating a fix in
unsegen).